backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Vector Database
pgvector==0.2.4

# AI/ML
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0

# Audio Processing
whisper==1.1.10

# Task Queue - Fixed version compatibility
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# Utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1